LPS-75031 Move implementation into new service
