+++ modules/apps/collaboration/document-library/document-library-sync-api/src/main/java/com/liferay/document/library/sync/constants/DLSyncConstants.java
@@ -12,14 +12,11 @@
  * details.
  */
 
-package com.liferay.document.library.kernel.model;
+package com.liferay.document.library.sync.constants;
 
 /**
  * @author Michael Young
- * @deprecated As of 7.0.0, replaced by {@link
- *             com.liferay.document.library.sync.constants.DLSyncConstants}
  */
-@Deprecated
 public class DLSyncConstants {
 
 	public static final String EVENT_ADD = "add";
