+++ modules/apps/collaboration/document-library/document-library-sync-service/src/main/java/com/liferay/document/library/sync/model/impl/DLSyncEventImpl.java
@@ -17,23 +17,8 @@ package com.liferay.document.library.sync.model.impl;
 import aQute.bnd.annotation.ProviderType;
 
 /**
- * The extended model implementation for the DLSyncEvent service. Represents a row in the &quot;DLSyncEvent&quot; database table, with each column mapped to a property of this class.
- *
- * <p>
- * Helper methods and all application logic should be put in this class. Whenever methods are added, rerun ServiceBuilder to copy their definitions into the {@link com.liferay.document.library.sync.model.DLSyncEvent} interface.
- * </p>
- *
  * @author Brian Wing Shun Chan
  */
 @ProviderType
 public class DLSyncEventImpl extends DLSyncEventBaseImpl {
-
-	/**
-	 * NOTE FOR DEVELOPERS:
-	 *
-	 * Never reference this class directly. All methods that expect a dl sync event model instance should use the {@link com.liferay.document.library.sync.model.DLSyncEvent} interface instead.
-	 */
-	public DLSyncEventImpl() {
-	}
-
 }
