+++ modules/apps/collaboration/document-library/document-library-sync-service/src/main/java/com/liferay/document/library/sync/service/impl/DLSyncEventLocalServiceImpl.java
@@ -14,28 +14,69 @@
 
 package com.liferay.document.library.sync.service.impl;
 
+import com.liferay.document.library.sync.model.DLSyncEvent;
 import com.liferay.document.library.sync.service.base.DLSyncEventLocalServiceBaseImpl;
+import com.liferay.portal.kernel.dao.orm.DynamicQuery;
+import com.liferay.portal.kernel.dao.orm.DynamicQueryFactoryUtil;
+import com.liferay.portal.kernel.dao.orm.Projection;
+import com.liferay.portal.kernel.dao.orm.ProjectionFactoryUtil;
+import com.liferay.portal.kernel.dao.orm.Property;
+import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
+
+import java.util.List;
 
 /**
- * The implementation of the dl sync event local service.
- *
- * <p>
- * All custom service methods should be put in this class. Whenever methods are added, rerun ServiceBuilder to copy their definitions into the {@link com.liferay.document.library.sync.service.DLSyncEventLocalService} interface.
- *
- * <p>
- * This is a local service. Methods of this service will not have security checks based on the propagated JAAS credentials because this service can only be accessed from within the same VM.
- * </p>
- *
- * @author Brian Wing Shun Chan
- * @see DLSyncEventLocalServiceBaseImpl
- * @see com.liferay.document.library.sync.service.DLSyncEventLocalServiceUtil
+ * @author Dennis Ju
  */
 public class DLSyncEventLocalServiceImpl
 	extends DLSyncEventLocalServiceBaseImpl {
 
-	/**
-	 * NOTE FOR DEVELOPERS:
-	 *
-	 * Never reference this class directly. Always use {@link com.liferay.document.library.sync.service.DLSyncEventLocalServiceUtil} to access the dl sync event local service.
-	 */
+	@Override
+	public DLSyncEvent addDLSyncEvent(String event, String type, long typePK) {
+		DLSyncEvent dlSyncEvent = dlSyncEventPersistence.fetchByTypePK(typePK);
+
+		if (dlSyncEvent == null) {
+			long dlSyncEventId = counterLocalService.increment();
+
+			dlSyncEvent = dlSyncEventPersistence.create(dlSyncEventId);
+
+			dlSyncEvent.setType(type);
+			dlSyncEvent.setTypePK(typePK);
+		}
+
+		dlSyncEvent.setModifiedTime(System.currentTimeMillis());
+		dlSyncEvent.setEvent(event);
+
+		return dlSyncEventPersistence.update(dlSyncEvent);
+	}
+
+	@Override
+	public void deleteDLSyncEvents() {
+		dlSyncEventPersistence.removeAll();
+	}
+
+	@Override
+	public List<DLSyncEvent> getDLSyncEvents(long modifiedTime) {
+		return dlSyncEventPersistence.findByModifiedTime(modifiedTime);
+	}
+
+	@Override
+	public List<DLSyncEvent> getLatestDLSyncEvents() {
+		DynamicQuery dynamicQuery = DynamicQueryFactoryUtil.forClass(
+			DLSyncEvent.class);
+
+		Property property = PropertyFactoryUtil.forName("modifiedTime");
+
+		DynamicQuery modifiedTimeDynamicQuery =
+			DynamicQueryFactoryUtil.forClass(DLSyncEvent.class);
+
+		Projection projection = ProjectionFactoryUtil.max("modifiedTime");
+
+		modifiedTimeDynamicQuery.setProjection(projection);
+
+		dynamicQuery.add(property.eq(modifiedTimeDynamicQuery));
+
+		return dlSyncEventPersistence.findWithDynamicQuery(dynamicQuery);
+	}
+
 }
