+++ portal-impl/src/com/liferay/portlet/documentlibrary/model/impl/DLSyncEventImpl.java
@@ -16,6 +16,9 @@ package com.liferay.portlet.documentlibrary.model.impl;
 
 /**
  * @author Brian Wing Shun Chan
+ * @deprecated As of 7.0.0, replaced by {@link
+ *             com.liferay.document.library.sync.model.impl.DLSyncEventImpl}
  */
+@Deprecated
 public class DLSyncEventImpl extends DLSyncEventBaseImpl {
 }
