+++ portal-impl/src/com/liferay/portlet/documentlibrary/service/impl/DLSyncEventLocalServiceImpl.java
@@ -15,68 +15,49 @@
 package com.liferay.portlet.documentlibrary.service.impl;
 
 import com.liferay.document.library.kernel.model.DLSyncEvent;
-import com.liferay.portal.kernel.dao.orm.DynamicQuery;
-import com.liferay.portal.kernel.dao.orm.DynamicQueryFactoryUtil;
-import com.liferay.portal.kernel.dao.orm.Projection;
-import com.liferay.portal.kernel.dao.orm.ProjectionFactoryUtil;
-import com.liferay.portal.kernel.dao.orm.Property;
-import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
 import com.liferay.portlet.documentlibrary.service.base.DLSyncEventLocalServiceBaseImpl;
 
 import java.util.List;
 
 /**
  * @author Dennis Ju
+ * @deprecated As of 7.0.0, replaced by {@link
+ *             com.liferay.document.library.sync.service.impl.DLSyncEventLocalServiceImpl}
  */
+@Deprecated
 public class DLSyncEventLocalServiceImpl
 	extends DLSyncEventLocalServiceBaseImpl {
 
 	@Override
 	public DLSyncEvent addDLSyncEvent(String event, String type, long typePK) {
-		DLSyncEvent dlSyncEvent = dlSyncEventPersistence.fetchByTypePK(typePK);
-
-		if (dlSyncEvent == null) {
-			long dlSyncEventId = counterLocalService.increment();
-
-			dlSyncEvent = dlSyncEventPersistence.create(dlSyncEventId);
-
-			dlSyncEvent.setType(type);
-			dlSyncEvent.setTypePK(typePK);
-		}
-
-		dlSyncEvent.setModifiedTime(System.currentTimeMillis());
-		dlSyncEvent.setEvent(event);
-
-		return dlSyncEventPersistence.update(dlSyncEvent);
+		throw new UnsupportedOperationException(
+			"This class is deprecated and replaced by " +
+				"com.liferay.document.library.sync.service.impl." +
+					"DLSyncEventLocalServiceImpl");
 	}
 
 	@Override
 	public void deleteDLSyncEvents() {
-		dlSyncEventPersistence.removeAll();
+		throw new UnsupportedOperationException(
+			"This class is deprecated and replaced by " +
+				"com.liferay.document.library.sync.service.impl." +
+					"DLSyncEventLocalServiceImpl");
 	}
 
 	@Override
 	public List<DLSyncEvent> getDLSyncEvents(long modifiedTime) {
-		return dlSyncEventPersistence.findByModifiedTime(modifiedTime);
+		throw new UnsupportedOperationException(
+			"This class is deprecated and replaced by " +
+				"com.liferay.document.library.sync.service.impl." +
+					"DLSyncEventLocalServiceImpl");
 	}
 
 	@Override
 	public List<DLSyncEvent> getLatestDLSyncEvents() {
-		DynamicQuery dynamicQuery = DynamicQueryFactoryUtil.forClass(
-			DLSyncEvent.class);
-
-		Property property = PropertyFactoryUtil.forName("modifiedTime");
-
-		DynamicQuery modifiedTimeDynamicQuery =
-			DynamicQueryFactoryUtil.forClass(DLSyncEvent.class);
-
-		Projection projection = ProjectionFactoryUtil.max("modifiedTime");
-
-		modifiedTimeDynamicQuery.setProjection(projection);
-
-		dynamicQuery.add(property.eq(modifiedTimeDynamicQuery));
-
-		return dlSyncEventPersistence.findWithDynamicQuery(dynamicQuery);
+		throw new UnsupportedOperationException(
+			"This class is deprecated and replaced by " +
+				"com.liferay.document.library.sync.service.impl." +
+					"DLSyncEventLocalServiceImpl");
 	}
 
 }
